+++ src/context/MetricsContext/utils.ts
@@ -1,87 +1,7 @@
 import { OrderType } from "domain/synthetics/orders";
 import { extractError, isUserRejectedActionError, TxError } from "lib/contracts/transactionErrors";
 import { MetricsContextType } from "./MetricsContext";
-
-export type MetricEventType = OrderEventType | "positionsListLoad.success";
-export type MetricData = OrderMetricData | OrderWsEventMetricData | PendingTxnErrorMetricData;
-
-export type OrderEventType = `${OrderMetricType}.${OrderStageType}`;
-export type OrderStageType = "submitted" | "sent" | "created" | "executed" | "cancelled" | "rejected" | "failed";
-
-export type OrderMetricType =
-  | SwapMetricParams["metricType"]
-  | IncreaseOrderMetricParams["metricType"]
-  | DecreaseOrderMetricParams["metricType"]
-  | EditCollateralMetricParams["metricType"]
-  | "unknownTxn";
-
-export type OrderMetricData =
-  | SwapMetricParams
-  | IncreaseOrderMetricParams
-  | DecreaseOrderMetricParams
-  | EditCollateralMetricParams;
-
-export type OrderWsEventMetricData = (SwapMetricParams | IncreaseOrderMetricParams | DecreaseOrderMetricParams) & {
-  key: string;
-  txnHash: string;
-};
-
-export type PendingTxnErrorMetricData = {
-  metricType: OrderMetricType;
-  txnHash: string;
-};
-
-export type SwapMetricParams = {
-  metricType: "swap" | "limitSwap";
-  account: string | undefined;
-  initialCollateralTokenAddress: string | undefined;
-  toTokenAddress: string | undefined;
-  initialCollateralDeltaAmount: bigint | undefined;
-  minOutputAmount: bigint | undefined;
-  swapPath: string[] | undefined;
-  executionFee: bigint | undefined;
-  allowedSlippage: number | undefined;
-  orderType: OrderType | undefined;
-};
-
-export type PositionOrderMetricParams = {
-  account: string | undefined;
-  referralCodeForTxn: string | undefined;
-  hasExistingPosition: boolean | undefined;
-  marketAddress: string | undefined;
-  initialCollateralTokenAddress: string | undefined;
-  initialCollateralDeltaAmount: bigint | undefined;
-  swapPath: string[] | undefined;
-  sizeDeltaUsd: bigint | undefined;
-  sizeDeltaInTokens: bigint | undefined;
-  triggerPrice: bigint | undefined;
-  acceptablePrice: bigint | undefined;
-  isLong: boolean | undefined;
-  orderType: OrderType | undefined;
-  executionFee: bigint | undefined;
-};
-
-export type IncreaseOrderMetricParams = {
-  metricType: "increasePosition" | "limitOrder";
-} & PositionOrderMetricParams;
-
-export type DecreaseOrderMetricParams = {
-  metricType: "decreasePosition" | "takeProfitOrder" | "stopLossOrder";
-  place: "tradeBox" | "positionSeller";
-  isFullClose: boolean | undefined;
-} & PositionOrderMetricParams;
-
-export type EditCollateralMetricParams = {
-  metricType: "depositCollateral" | "withdrawCollateral";
-  account: string | undefined;
-  marketAddress: string | undefined;
-  initialCollateralTokenAddress: string | undefined;
-  initialCollateralDeltaAmount: bigint | undefined;
-  swapPath: [];
-  isLong: boolean | undefined;
-  orderType: OrderType | undefined;
-  executionFee: bigint | undefined;
-};
+import { OrderMetricType } from "./types";
 
 export function getMetricTypeByOrderType(p: {
   orderType: OrderType;
